Loop through pkt buf as array (not linked list)
When wrapping the buffers of the `send_pkt_buf` and `recv_pkt_buf`
arrays, loop through them as arrays rather than as a linked list.
Looping through as a linked list works immediately after initialization
(i.e. before any packets have been sent or received), but after that it
is possible for the linked list to be incomplete and/or no longer in
sequential order.