Drop use of CBinding in favor of direct ccall use
`CBinding` is convenient, but it imposes a layer of abstraction that has
an adverse impact on performance.  `HashpipeIBVerbs` is now implemented
using manually crafted structs and functions that use the `@ccall` macro
to call directly into the `hashpipe_ibverbs` library.  The
implementation is very allocation conscious and has extensive support
for navigating the underlying data structures using `Ptr` objects.
Performance is greatly improved over the CBinding approach.