Add examples/dibas_pktsink.jl
This script sinks DIBAS packets from the network into memory (and
ultimately discards them).  The dibas_pktgen.jl script can be used to
generate DIBAS packets and send them over the network.  This script is
currently setup to receive 100e6 packets and report stats every 10e6
packets.  The DIBAS mcount is extracted from the packets and used to
monitor for packet loss (and/or out of order packet reception) when
reporting stats.  This shows up as a Δmcount value that should be 0 if
no packet loss or reordering has occurred.  This currently
assumes/requires that the dibas_pktgen.jl script sends the first packet
after this script is ready to receive packets.